requirement.txt
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
pycparser==2.21
PySocks==1.7.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
allure-behave==2.11.1
allure-python-commons==2.11.1
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
